fix to_ascii under 32-bit msvc builds
Summary:
Under gcc and clang, we use the intrinsic `__builtin_clzll`. This builtin is available regardless of whether the `clz` instruction is available on the target platform - when it is, the backend chooses it; when it is not, the backend emits emulation code.

Under msvc, the builtin is `__lzcnt64`. However, this builtin is available only in x64 builds; there is no fallback to emulation code, and 32-bit builds fail to compile.

Differential Revision: D26918251

fbshipit-source-id: a4d9c199f632c5a210bd6fd8782d327f7aac774e